Fold the signal-hook-sys inside the main signal-hook
To avoid problems with extra dependency, compatibility, etc.
