+++ .github/workflows/test.yaml
@@ -151,6 +151,7 @@ jobs:
       - name: Check compilation
         run: |
           rm Cargo.lock
+          cargo update
           cargo check --no-default-features
