+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Allow extracting Origin from the raw `siginfo_t` structure by hand, without
   needing an iterator.
+* Folding the signal-hook-sys inline (but still compiling C code only
+  conditionally).
 
 # 0.3.1
