+++ Cargo.toml
@@ -20,7 +20,8 @@ maintenance = { status = "actively-developed" }
 channel = []
 default = ["channel", "iterator"]
 iterator = []
-extended-siginfo = ["channel", "iterator", "signal-hook-sys"]
+extended-siginfo = ["channel", "iterator", "extended-siginfo-raw"]
+extended-siginfo-raw = ["cc"]
 
 [workspace]
 members = [
@@ -29,16 +30,18 @@ members = [
     "signal-hook-tokio",
     "signal-hook-mio",
     "signal-hook-async-std",
-    "signal-hook-sys",
 ]
 
 [dependencies]
 libc = "^0.2"
 signal-hook-registry = { version = "^1.3", path = "signal-hook-registry" }
-signal-hook-sys = { version = "^0.1.1", path = "signal-hook-sys", optional = true }
 
 [dev-dependencies]
 serial_test = "^0.5"
 
 [package.metadata.docs.rs]
 all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
+[build-dependencies]
+cc = { version = "^1", optional = true }
