+++ signal-hook-async-std/Cargo.toml
@@ -30,3 +30,4 @@ serial_test = "~0.5"
 
 [package.metadata.docs.rs]
 all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
