--- signal-hook-sys/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "signal-hook-sys"
-version = "0.1.1"
-authors = ["Michal 'vorner' Vaner <vorner@vorner.cz>"]
-edition = "2018"
-description = "Backend crate for signal-hook, with some low-level C code"
-documentation = "https://docs.rs/signal-hook-sys"
-license = "Apache-2.0/MIT"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-libc = "~0.2"
-
-[build-dependencies]
-cc = "~1"
