--- signal-hook-sys/build.rs
@@ -1,5 +1,0 @@
-use cc::Build;
-
-fn main() {
-    Build::new().file("src/extract.c").compile("extract");
-}
