+++ signal-hook-tokio/Cargo.toml
@@ -37,3 +37,4 @@ serial_test = "~0.5"
 
 [package.metadata.docs.rs]
 all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
