+++ src/low_level/mod.rs
@@ -10,7 +10,7 @@ use libc::c_int;
 pub mod channel;
 #[cfg(not(windows))]
 pub mod pipe;
-#[cfg(feature = "signal-hook-sys")]
+#[cfg(feature = "extended-siginfo-raw")]
 pub mod siginfo;
 
 pub use signal_hook_registry::{register, unregister};
